zahtjev adding, zahtjev and zadatak view

+++ frontend/DEVOPSTRK/src/components/Navbar.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="flex flex-col bg-cyan-900 h-screen justify-between w-64 shadow">
+    <div class="flex flex-col bg-cyan-900 h-auto sm:h-screen justify-between w-64 shadow flex-shrink-0">
         <div class="flex items-center text-3x1 px-5 p-5 bg-cyan-800 shadow">
             <img src="devops.png" class="w-10 h-10">
             <p class="mx-3 text-2xl font-bold">DevOpsTrk</p>
@@ -75,9 +75,11 @@ export default {
     },
     methods: {
         async logout() {
-            this.$store.state.user = undefined
-            sessionStorage.clear()
-            this.$router.push("/Login")
+            if (confirm("Odjava iz sustava?")) {
+                this.$store.state.user = undefined
+                sessionStorage.clear()
+                this.$router.push("/Login")
+            }
         },
         storeInSession(value) {
             sessionStorage.setItem("projektInfoPage", value)
